Sped up the ETL handler.
Main performance gain was by minimizing calls to the database.
Previously for each word in every tweet, the application looked in
memcache, and then the database. In a batch of 20 tweets, even at 10
words per tweet, it meant 200 memcache/database calls.

This patch collects all the words into a dictionary mapping words to
tweets. It then makes one call to the database to fetch corresponding
topics, significantly speeding up the ETL process.