This is an optimization to address the issue of the ETL task using about 20 seconds of CPU per run. If the task is run every minute, the CPU limit of 6.5 hours is reached in a day.
The optimization requires topic names be stored as trees of words. For
example, if a user creates the topic "easter egg", we now store a topic
for each word:

easter = Topic(key_name='parent:easter', name='easter')
egg = Topic(key_name='key:egg', parent=easter, name='easter egg')

Previously, we would store a single topic:

Topic(key_name='key:easter egg', name='easter egg')

The tree structure of topics in the datastore allows the application to
retrieve topics beginning with different words in a single query. This
is not possible with a model's properties.

Previously, to retrieve topics beginning with 'easter', 'snakes', or
'plane', the application would have to make multiple queries.

Topic.gql("WHERE name > :1 AND name < :2", \
          'easter', u'easter' + u'\ufffd')
Topic.gql("WHERE name > :1 AND name < :2", \
          'snake', u'snake' + u'\ufffd')
Topic.gql("WHERE name > :1 AND name < :2", \
          'plane', u'plane' + u'\ufffd')

With the current optimization, the application can achieve the same
result in a single query.

Topic.get_by_key_name(['easter', 'snakes', 'plane']

The new ETL process makes a single query and uses the result of that
query to filter topics. To give an idea of quantities, each tweet is 140
characters; and, because we split on spaces, the maximum word count is 70.
The number of possible topics 70 words can create is (70 * (70 + 1)) /
2 = 2485.

This patch introduces the following optimization:
1. Each tweet is split into words, and then sent as key names to the
   database. One query for parents, and another query for singles.
2. The database returns a same sized list, with Topics in place of the key
   names, and None where no Topic with that key name exists.
3. We then iterate over the returned parent Topics, and create paths. These
   paths are sets of phrases that begin with the parent topic and end with
   the child of each generation.
4. These paths are turned into keys and sent to the database. The database
   replies with a list of Topics and Nones. We iterate over this list and
   match the topics to the Tweet.
5. Finally, we save the Tweets which have Topics matched.  Removed unused
   functions.

The average number of keys was about 3000, most of them returning None.
It is now 2 queries of about 200 keys, then a final single query which
depends on the number of parent topics in the datastore.

The effectiveness of this optimization relies on the number of parent
topics in the datastore.