Redesigned the experience of the site with a new look and making creation of new topics simpler.
Changes include:

A more colorful theme, and slicker logo with a visible tagline. May have
went too far on the rounded-corners, oh well.

Removed the confirmation step to create a topic. When edrop encountered a
new topic, it used to ask if the user wanted to continue creating it.
Now the topic is immediately created.