Bug fix and test for issue with multiword topics whose parent nodes exist as in the database.
For example, the topics "of montreal" and "time of day" share the same
node "of", which has a key name "parent:of". Entities are unique per
key/parent pair, therefore there may be more than one entity with key
name "parent:of".

If the following node chain exists in the datastore

  "parent:of" => "key:montreal"

the subsequent creation of

  "parent:time" => "parent:of" => "key:day"

would have used the existing "of" even though it should create a brand
new topic with the parent "parent:time". Topics with this flaw should be
fixed in the datastore.